[af-19] feat: MemberrRole Convertor 적용
