+++ src/test/java/com/drunkenlion/alcoholfriday/domain/admin/member/api/AdminMemberControllerTest.java
@@ -3,7 +3,6 @@ package com.drunkenlion.alcoholfriday.domain.admin.member.api;
 import com.drunkenlion.alcoholfriday.domain.member.dao.MemberRepository;
 import com.drunkenlion.alcoholfriday.domain.member.entity.Member;
 import com.drunkenlion.alcoholfriday.domain.auth.enumerated.ProviderType;
-import com.drunkenlion.alcoholfriday.domain.member.enumerated.MemberRole;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
