+++ src/test/java/com/drunkenlion/alcoholfriday/domain/admin/member/application/AdminMemberServiceTest.java
@@ -6,7 +6,6 @@ import com.drunkenlion.alcoholfriday.domain.admin.member.dto.MemberModifyRequest
 import com.drunkenlion.alcoholfriday.domain.member.dao.MemberRepository;
 import com.drunkenlion.alcoholfriday.domain.member.entity.Member;
 import com.drunkenlion.alcoholfriday.domain.auth.enumerated.ProviderType;
-import com.drunkenlion.alcoholfriday.domain.member.enumerated.MemberRole;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
